debian: Fix restarting problem when upgrading from non-split.
Use dh_installinit -R to restart in the postinst. This has two
benefits: it will fix the restarting problem from non-split packages,
and it will also minimize the downtime syslog-ng is stopped.

This is how the non-split Debian package works too, it just does this
a in a more complicated manner.