debian: Use dpkg triggers.
Since we want to restart syslog-ng when a module package is changed,
and we don't want to litter their maintainer scripts with duplicated
code, use dpkg triggers instead.

This also means that we need to restart syslog-ng less (twice, during
upgrades: once when syslog-ng-core is configured, and one more time
after all the modules are done).