debian: Add full ubuntu lucid support.
As ubuntu lucid does not support data.tar.xz, and I want to reuse the
arch:all packages on these systems too, it needs to be compressed
without xz. The easiest way to accomplish this is to always build the
binary-indep package without xz, even on distributions that do support
data.tar.xz.

The rest is built with xz on every distribution but lucid, where it is
built without.