syslog-ng: Add a -q/--silent-module-load option
With the -q/--silent-module-load option set (defaults to off), no
warnings will be printed when a module cannot be found.

If it's found, but cannot be loaded, a message is still printed. If it
is used later in the config, that's still handled correctly.

The flag only suppresses the warning, and nothing else.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>