[logproto.c] "Invalid Frame Header" in syslog protocol
When using the syslog protocol, in case when we received an EAGAIN
during an attempt to send the framing, the frame was not resent,
resulting in protocol errors and invalid frame header messages on the
other end.

This patch corrects the problem by not using
log_proto_text_client_post to send framed messages, but a wrapper that
can resume sending the frame too.

Signed-off-by: Fried Zoltan <deirf@balabit.hu>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>