LogThreadedDestDriver: A new threaded destination driver base class
This implements a new class, LogThreadedDestDriver, whose goal is to
be the base class for all destination drivers that use a worker thread
to do their blocking work.

For now, the class provides enough functionality to move most - if not
all - of the boring thread and queue handling stuff out of all drivers
whose worker thread handles a single message at a time (such as
afmongodb and afsmtp).

The intended use is to make this the parent of the destination driver,
instead of LogDestDriver, and implement the various callbacks:

* worker_job_method: The central point of it all, the function that
  does the actual work.
* worker_job_error: The function to call when an error occurs during
  worker_job_method. The whole driver will be suspended for
  time_reopen time afterwards.
* worker_thread_init_method: Called once, to initialize the worker
  thread. This function can be used to allocate memory for whatever
  structures the thread needs, and perform other kinds of setup, such
  as connecting to a database.
* worker_thread_deinit_method: The counterpart of
  worker_thread_init_method: this one is for deinitializing the
  driver.
* worker_thread_method: The worker thread function itself. The class
  provides a default, which should work for the vast majority of
  cases.
* worker_thread_job_method: Called from the default
  worker_thread_method: this function does the queue handling magic,
  and in turn, calls worker_job_method.

  The driver comes with a reasonable default.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>