redisdd: A simple Redis destination driver
This is an implementation of a very simple Redis destination
driver. It does not use the async API of hiredis, and is made for
syslog-ng 3.3 on request, to fulfill a particular need and pretty much
only that.

The driver is able to connect to redis and send commands to it, where
the command always takes the form of "$CMD $PREFIX$KEY $VALUE". The
first two is configurable via the command() and prefix() settings, and
default to "LPUSH" and an empty string by default. The latter two are
the key-value pairs selected by the value-pairs setting, which
defaults to selected-macros and nv-pairs.

In the future, this will be ported to 3.4, using the
LogThreadedDestDriver class (or the async API of hiredis), and support
for MULTI will be added too. At that point, the prefix() setting will
vanish, because in syslog-ng 3.4 the same can be accomplished via
value-pairs transformations.

Example usage, with all options show-cased:

destination d_redis {
 redis(host("localhost") port(6379)
       command("RPUSH") prefix("syslog-ng.")
       value-pairs(key("MESSAGE") pair("DATE" "$R_FULLDATE")));
};

Requested-by: Ferenc Kalcso <ferenc.kalcso@duodecad.hu>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>