afmongodb: Add support for datetime fields
For TTL collections primarly (but useful for other scenarios aswell),
introduce a new option: datetime(FIELD_NAME). This allows one to mark
a single field anywhere in the structure as a datetime field, and the
driver will attempt to convert it to a proper datetime, by assuming it
is a unix timestamp.

If the conversion fails, or the field is not found, it will not be
inserted.

Usage:

mongodb(datetime("some.date")
        value-pairs(
          key("MESSAGE")
          pair("some.date", "$UNIXTIME")))

This will result in something like:

{ "_id" : ObjectId("509b8ffc1a1006213a000001"),
  "some" : { "date" : ISODate("2012-11-08T10:56:59Z") },
  "MESSAGE" : "foobar" }

Signed-off-by: Gergely Nagy <algernon@balabit.hu>