luaconfig: Added preliminary version of lua destination.
This is just a proof-of-concept version of lua destination, which
takes a function name at config time, copies the function's byte-code to
a clean new lua state, and calls this function every time a log message
arrives at the lua destination. The function should have only one parameter,
which is the formatted version of the LogMessage by the template "${MESSAGE}".
(in other words: the value of the MESSAGE macro for that log message).

Example:

function test_dest_function(msg)
   print(msg)
end

Destination("d_lua", { LuaDest("test_dest_function") })