stomp: Removed routing key from stomp driver.
Because routing_key is an amqp idiom, in stomp this is called
destination. It is used only in stomp, when one wants to interop
with amqp, but in this case, it is rarely used on the publisher side.
If somebody wants to use it, it can be achieved by using a rewrite rule,
setting routing_key there, and using value-pairs in the stomp driver.